Intel(R) SDP Tool Action plugin
Intel(R) SDP Tool Ansible plugin is an Ansible action plugin, which provides
Intel(R) Server Debug and Provisioning Tool (SDP Tool) capabilities in an
action plugin. The action plugin provides for all the capabilities of the
Intel(R) SDP Tool. Refer to the documentation of Intel(R) SDP Tool for more
details on usage, command list and arguments.

Tests : Unit tests and intergration tests are available in the source.

Samples and Examples : Samples and examples are also available in the playbook
directory.